[sui-metrics-proxy/ initial release] # Summary:
We wish to eliminate using third party metric agents for collecting metrics from sui nodes.  This pr creates a metric proxy service that accepts post data from downstream nodes and relays it to an upstream tsdb.

The design of the proxy is intended to return a success to calling clients as early as possible.  It does this by queueing up the client data into an mpsc channel and then that will be sent upstream.  After sending on the channel, (which is buffered), the client will get an immediate "accept".  This status indicates that we have accepted the metrics and intend to relay them upstream.  Some failure modes here are restarts (we drop whatever we have queued on the floor) or if we can't relay them upstream correctly, they will be dropped.  Future work can optimize this for stateful recovery if we find it necessary.

todo in future pr:
* relay data to a tsdb data sink
* add more tests
* decide what max post data size will be

# Test Plan:

cargo test and interactive testing

### interactive:

to start the server:
```
cargo run -- -l="[::1]:8080" -u="[::1]:8081" -b 100
```

use tests to simulate client posts.